requirements.txt
pandas==2.3.0
plotly==6.1.2
dash==3.0.4
yfinance==0.2.63
Flask-Caching==2.3.1
cachetools==6.0.0
scipy==1.15.3
dash-bootstrap-components==2.0.3
gunicorn==23.0.0
APScheduler==3.11.0
requests==2.32.4
exchange_calendars==4.10.1
dateparser==1.2.1
orjson==3.10.18
python-dotenv==1.1.0
numba==0.61.2
flask==3.0.1
discord.py==2.5.0
aiofiles==24.1.0
python-dotenv==1.1.0
matplotlib==3.10.3